// src/views/Advice/types.ts
export interface IFund {
  fundId: string;
  fundNm: string;
  price: string;
  volume: string;
  discountRt: string;
  applyFee: string;
  applyFeeTips: string;
  redeemFee: string;
  redeemFeeTips: string;
  estimateValue: string;
  estimateDate: string;
  applyStatus: string;
  redeemStatus: string;
}